Plugin updated for godot 4.3. Plugin structure and dependencies updated to latest versions.
