+++ LICENSE
@@ -1,6 +1,7 @@
 MIT License
 
 Copyright (c) 2022 WolfBearGames
+Copyright 2024 Gabriel Olekšák (www.nevadza.com)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
