--- app/src/main/AndroidManifest.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="de.wolfbeargames.geolocationplugin">
-
-    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-
-    <application>
-        <meta-data
-            android:name="org.godotengine.plugin.v1.GeolocationPlugin"
-            android:value="de.wolfbeargames.geolocationplugin.GeolocationPlugin" />
-    </application>
-
-</manifest>
