--- build.gradle
@@ -1,10 +1,0 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-plugins {
-    id 'com.android.application' version '7.2.2' apply false
-    id 'com.android.library' version '7.2.2' apply false
-    id 'org.jetbrains.kotlin.android' version '1.6.10' apply false
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
-}
