+++ gradle/libs.versions.toml
@@ -1,0 +1,12 @@
+[versions]
+agp = "8.5.2"
+kotlin = "2.0.20"
+gms = "21.3.0"
+
+[libraries]
+google-gms = {group = "com.google.android.gms", name = "play-services-location", version.ref = "gms"}
+
+[plugins]
+android-library = { id = "com.android.library", version.ref = "agp" }
+jetbrains-kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
+
